agents/requirements.txt
agentpy==0.1.5
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asciitree==0.3.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
azure-ai-agents==1.1.0b2
azure-ai-projects==1.0.0b5
azure-core==1.32.0
azure-cosmos==4.9.0
azure-identity==1.20.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.10.0
azure-keyvault-keys==4.11.0
azure-keyvault-secrets==4.10.0
azure-storage-blob==12.25.1
blinker==1.9.0
branca==0.8.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
commonmark==0.9.1
contourpy==1.3.0
cryptography==44.0.1
cssselect2==0.8.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fasteners==0.19
fiona==1.10.1
folium==0.20.0
fonttools==4.58.4
frozenlist==1.7.0
geographiclib==2.0
geopandas==0.14.4
geopy==2.4.1
geotiff==0.2.10
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imagecodecs==2023.9.18
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonlines==3.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
landfire==0.5.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
msal==1.31.1
msal-extensions==1.2.0
multidict==6.4.4
multiprocess==0.70.18
narwhals==1.43.0
nest-asyncio==1.6.0
networkx==3.2.1
noise==1.2.2
numcodecs==0.12.1
numpy==1.26.4
openai==1.88.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.3.0
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.8
plotly==6.1.2
portalocker==2.10.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyWavelets==1.6.0
pywin32==308
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
reportlab==3.6.13
requests==2.32.3
rich==12.6.0
rpds-py==0.25.1
SALib==1.5.1
scikit-image==0.19.3
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.7
simfire==2.0.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.1
streamlit-option-menu==0.4.0
streamlit_folium==0.25.0
svglib==1.5.1
tenacity==9.1.2
tifffile==2024.8.30
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.3
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
wurlitzer==3.1.1
xyzservices==2025.4.0
yarl==1.20.1
zarr==2.12.0
zipp==3.23.0
semantic-kernel